Extract createPipelineFromOptions
